requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==5.0.0
boto3==1.40.59
botocore==1.40.59
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg-pool==3.2.6
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
